OPENDJ-1177 (CR-3314) Re-implement changelog purging logic
JEChangelogDB.java:
In ChangelogDBPurger.run(), changed code to:
- support the absence of change number index DB. To simplify matters, code assumes that ds-cfg-compute-change-number does not change during the life time of an RS.
- purged by using a CSN made up from the purge delay rather than using the previous cookie.
- sleep 500 millis if there are no changes to purge, or sleep till the next change to purge.
- gracefully shutdown without fuss in the logs.


JEChangeNumberIndexDB.java
In purgeUpTo(), return the oldest non purged CSN rather than the previous cookie + merged two branches of the code.

ExternalChangeLogTest.java:
Consequence of the change to JEChangeNumberIndexDB.
Extracted method assertECLLimits() from ECLCompatTestLimits() + added a loop inside it to let the code persist changes asynchronously from the test thread.

JEChangeNumberIndexDBTest.java:
Renamed testTrim() to testPurge().
In newReplicationServer(), enabled ds-cfg-compute-change-number.

+++ opendj3-server-dev/src/server/org/opends/server/replication/server/changelog/je/JEChangeNumberIndexDB.java
@@ -37,7 +37,6 @@ import org.opends.server.admin.std.server.MonitorProviderCfg;
 import org.opends.server.api.MonitorProvider;
 import org.opends.server.core.DirectoryServer;
 import org.opends.server.replication.common.CSN;
-import org.opends.server.replication.common.MultiDomainServerState;
 import org.opends.server.replication.server.changelog.api.*;
 import org.opends.server.replication.server.changelog.je.DraftCNDB.*;
 import org.opends.server.types.*;
@@ -212,22 +211,18 @@ public class JEChangeNumberIndexDB implements ChangeNumberIndexDB
    * Synchronously purges the change number index DB up to and excluding the
    * provided timestamp.
    *
-   * @param purgeTimestamp
+   * @param purgeCSN
    *          the timestamp up to which purging must happen
-   * @return the {@link MultiDomainServerState} object that drives purging the
-   *         replicaDBs.
+   * @return the oldest non purged CSN.
    * @throws ChangelogException
    *           if a database problem occurs.
    */
-  public MultiDomainServerState purgeUpTo(long purgeTimestamp)
-      throws ChangelogException
+  public CSN purgeUpTo(CSN purgeCSN) throws ChangelogException
   {
-    if (isEmpty())
+    if (isEmpty() || purgeCSN == null)
     {
       return null;
     }
-
-    final CSN purgeCSN = new CSN(purgeTimestamp, 0, 0);
 
     final DraftCNDBCursor cursor = db.openDeleteCursor();
     try
@@ -239,21 +234,18 @@ public class JEChangeNumberIndexDB implements ChangeNumberIndexDB
         {
           oldestChangeNumber = record.getChangeNumber();
         }
-        if (record.getChangeNumber() == newestChangeNumber)
-        {
-          // do not purge the newest record to avoid having the last generated
-          // changenumber dropping back to 0 if the server restarts
-          return getPurgeCookie(record);
-        }
 
-        if (record.getCSN().isOlderThan(purgeCSN))
+        if (record.getChangeNumber() != newestChangeNumber
+            && record.getCSN().isOlderThan(purgeCSN))
         {
           cursor.delete();
         }
         else
         {
-          // Current record is not old enough to purge.
-          return getPurgeCookie(record);
+          // 1- Current record is not old enough to purge.
+          // 2- Do not purge the newest record to avoid having the last
+          // generated changenumber dropping back to 0 when the server restarts
+          return record.getCSN();
         }
       }
 
@@ -273,13 +265,6 @@ public class JEChangeNumberIndexDB implements ChangeNumberIndexDB
     {
       cursor.close();
     }
-  }
-
-  private MultiDomainServerState getPurgeCookie(
-      final ChangeNumberIndexRecord record) throws DirectoryException
-  {
-    // Do not include the record's CSN to avoid having it purged
-    return new MultiDomainServerState(record.getPreviousCookie());
   }
 
   /**
